requirements.txt
asgiref==3.8.1
boto3==1.34.113
botocore==1.34.113
dj-database-url==2.1.0
Django==5.0.6
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.9
django-appconf==1.0.6
django-ckeditor==6.7.1
django-compressor==4.4
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-js-asset==2.2.0
django-sass-processor==1.4
django-storages==1.14.3
jmespath==1.0.1
pillow==10.3.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
rcssmin==1.1.1
rjsmin==1.2.1
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
urllib3==2.2.1
whitenoise==6.6.0
